requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2021.10.8
    # via skyfield
click==8.1.2
    # via pip-tools
hijri-converter==2.2.3
    # via -r requirements.in
jplephem==2.17
    # via skyfield
numpy==1.22.3
    # via
    #   jplephem
    #   skyfield
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2022.1
    # via -r requirements.in
sgp4==2.21
    # via skyfield
six==1.16.0
    # via python-dateutil
skyfield==1.42
    # via -r requirements.in
tomli==2.0.1
    # via pep517
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
